// examples/workflow/main/src/lib/model/enum/actions/start-workflow-action-params.enum.ts
/*
 * Copyright(c) RIB Software GmbH
 */

/**
 * Input/Output parameters for the action 'Start Workflow Action'
 */
export enum StartWorkflowActionParamsEnum {
	EntityId = 'EntityId',
	StartContext = 'StartContext',
	WorkflowId = 'WorkflowId',
	ResultContext = 'ResultContext',
	CallingInstanceId = 'CallingInstanceId',
	Hours2Live = 'Hours2Live'
}